+++ widgets/views/wallEntry.php
@@ -2,28 +2,23 @@
 
 use humhub\modules\wiki\assets\Assets;
 use humhub\modules\wiki\helpers\Url;
-use humhub\modules\wiki\widgets\WikiRichText;
+use humhub\modules\wiki\models\WikiPage;
 use humhub\widgets\Button;
 
-/* @var $wiki \humhub\modules\wiki\models\WikiPage */
+/* @var $wiki WikiPage */
 /* @var $content string */
 
 Assets::register($this);
-
-$wikiUrl = Url::toWiki($wiki);
-
 ?>
 <div>
     <div class="wiki-preview">
         <div class="wiki-preview-content" data-ui-markdown>
-            <?php if(!empty($content)) : ?>
-                <?= WikiRichText::preview($content, 500, ['exclude' => ['anchor']]) ?>
-            <?php else: ?>
-                <?= Yii::t('WikiModule.base', 'This page is empty.') ?>
-            <?php endif; ?>
+            <?= empty($content)
+                ? Yii::t('WikiModule.base', 'This page is empty.')
+                : $content ?>
         </div>
 
-        <?= Button::asLink(Yii::t('UiModule.base', 'Read more'), $wikiUrl) ?>
+        <?= Button::asLink(Yii::t('UiModule.base', 'Read more'), Url::toWiki($wiki)) ?>
     </div>
 
 </div>